Move detectors to a separate module
